+++ src/Module/ModuleFactory.php
@@ -13,21 +13,6 @@ class ModuleFactory implements ModuleFactoryAlias
 {
 
     /**
-     * Holds a module builder to build the module
-     * 
-     * @var ModuleBuilderContract
-     */
-    private $moduleBuilder;
-
-    /**
-     * @param ModuleBuilderContract $moduleBuilder Builder to create the module object
-     */
-    public function __construct(ModuleBuilderContract $moduleBuilder)
-    {
-        $this->moduleBuilder = $moduleBuilder;
-    }
-
-    /**
      * Create a module from its alias
      * 
      * @param string $alias Alias of the module
@@ -35,15 +20,16 @@ class ModuleFactory implements ModuleFactoryAlias
      */
     public function fromAlias(string $alias): Module
     {
-        $this->moduleBuilder->create($alias);
-        $this->moduleBuilder->setAlias();
-        $this->moduleBuilder->setName();
-        $this->moduleBuilder->setDescription();
-        $this->moduleBuilder->setPermissions();
-        $this->moduleBuilder->setSettings();
-        $this->moduleBuilder->setTriggers();
-        $this->moduleBuilder->setCompletionConditions();
-        $this->moduleBuilder->setServices();
-        return $this->moduleBuilder->getModule();
+        $moduleBuilder = app(ModuleBuilderContract::class);
+        $moduleBuilder->create($alias);
+        $moduleBuilder->setAlias();
+        $moduleBuilder->setName();
+        $moduleBuilder->setDescription();
+        $moduleBuilder->setPermissions();
+        $moduleBuilder->setSettings();
+        $moduleBuilder->setTriggers();
+        $moduleBuilder->setCompletionConditions();
+        $moduleBuilder->setServices();
+        return $moduleBuilder->getModule();
     }
 }