+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [1.0.1] - (23/01/2020)
+
+### Changed
+- Generate a new module builder each time the module factory is used
+- Make module instance setting values nullable 
+- Connection ID attribute is now fillable for a module instance service
+- Prefer stable dependencies
+
 ## [1.0] - (17/01/2020)
 
 ### Added
 - Initial Release
 
-[unreleased]: https://github.com/bristol-su/support/compare/v1.0...HEAD
-[1.0]: https://github.com/bristol-su/support/releases/tag/v0.1...v1.0
-[0.1]: https://github.com/bristol-su/support/releases/tag/v0.1
+[Unreleased]: https://github.com/bristol-su/support/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/bristol-su/support/releases/tag/v1.0...v1.0.1
+[1.0]: https://github.com/bristol-su/support/releases/tag/v1.0